refactor: separate user snap test and move to position 10
